requirements.txt
attrs==25.4.0
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pystalk==0.7.0
python-dotenv==1.2.1
PyYAML==6.0.3
ruff==0.15.0
six==1.17.0
Werkzeug==3.1.5
